+++ lib/collection.dart
@@ -1,16 +1,17 @@
 /// A collection of new collection types: bi-map, bit-list, multi-set, set and
 /// list multi-map, range, and string.
-export 'src/collection/bigint_range.dart'
-    show BigIntRange, BigIntRangeExtension;
 export 'src/collection/bimap.dart' show BiMap, BiMapExtension;
 export 'src/collection/bitlist.dart' show BitList, BitListExtension;
-export 'src/collection/double_range.dart'
-    show DoubleRange, DoubleRangeExtension;
-export 'src/collection/integer_range.dart'
-    show IntegerRange, IntegerRangeExtension;
 export 'src/collection/multimap/list.dart' show ListMultimap;
 export 'src/collection/multimap/set.dart' show SetMultimap;
 export 'src/collection/multiset.dart' show Multiset, MultisetExtension;
+export 'src/collection/range.dart' show Range;
+export 'src/collection/range/bigint_range.dart'
+    show BigIntRange, BigIntRangeExtension;
+export 'src/collection/range/double_range.dart'
+    show DoubleRange, DoubleRangeExtension;
+export 'src/collection/range/integer_range.dart'
+    show IntegerRange, IntegerRangeExtension, IndicesIterableExtension;
 export 'src/collection/string.dart' show ListExtension;
 export 'src/collection/trie.dart'
     show Trie, TrieNode, TrieNodeEntry, TrieNodeList, TrieNodeMap;